requirements.txt
Django==5.1
django-unfold==0.19.0
environs==10.0.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
drf-yasg==1.21.7
django-filter==24.3
django-apscheduler==0.6.2
APScheduler==3.10.4
beautifulsoup4==4.12.3
selenium==4.23.1
unidecode==1.3.8
pillow==10.4.0
